ServerlessComputing/GoogleCloud/src_client/requirements.txt
request
google-auth==1.8.2